% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_reshape.R
\name{check_for_units}
\alias{check_for_units}
\title{Check for unique units per parameter}
\usage{
check_for_units(df_in)
}
\arguments{
\item{df_in}{Data frame loaded with \code{\link[=load_lab_data]{load_lab_data()}}}
}
\value{
Returns either a warning or a confirmation
}
\description{
Check for unique units per parameter
}
