% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs_functions.R
\name{discard_na_samples}
\alias{discard_na_samples}
\title{Removes all samples without complete tracer data}
\usage{
discard_na_samples(MCS_input)
}
\arguments{
\item{MCS_input}{A list of tracer concentrations per sample}
}
\value{
A list of tracer concentrations per sample
}
\description{
Removes all samples without complete tracer data
}
