% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs_functions.R
\name{get_all_sample_concentrations}
\alias{get_all_sample_concentrations}
\title{Random concentrations in all samples based on measured data and assumed
deviation}
\usage{
get_all_sample_concentrations(
  input_table,
  predictor_samples,
  tracer_names,
  MCS_runs,
  rel_deviation
)
}
\arguments{
\item{input_table}{Data frame reshaped by function \code{\link[=reshape_table]{reshape_table()}}}

\item{predictor_samples}{A character vector of names of samples used
for predicting the substance concentration of the remaining samples.}

\item{tracer_names}{A vector of character strings defining the tracers used}

\item{MCS_runs}{The number of random samples drawn
(Monte-Carlo-Simulation runs)}

\item{rel_deviation}{Either one numeric value between 0 and 1 defining the
standard deviation relative to the measured value for all tracers or a vector
of the same length as 'tracer_names' defining a deviation for each tracer}
}
\value{
A List of data frames with columns corresponding to the tracers and
rows correspoding to the Monte Carlo runs. The predictor samples are the
first three entries and are named "A", "B" and "C"
}
\description{
Random concentrations in all samples based on measured data and assumed
deviation
}
