% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{get_concentration_ranges}
\alias{get_concentration_ranges}
\title{Find minimum and maximum of the tracers in all samples (including MCS)}
\usage{
get_concentration_ranges(MCS_input)
}
\arguments{
\item{MCS_input}{A list of sample Data frames created by
\code{\link[=get_all_sample_concentrations]{get_all_sample_concentrations()}}}
}
\description{
Find minimum and maximum of the tracers in all samples (including MCS)
}
