% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcs_functions.R
\name{get_varrying_sample_concentration}
\alias{get_varrying_sample_concentration}
\title{Random concentrations in sample based on measured data and assumed deviation}
\usage{
get_varrying_sample_concentration(
  input_table,
  sample_name,
  tracer_names,
  MCS_runs,
  rel_deviation
)
}
\arguments{
\item{input_table}{Data frame reshaped by function \code{\link[=reshape_table]{reshape_table()}}}

\item{sample_name}{A character string defining the sample}

\item{tracer_names}{A vector of character strings defining the tracers used}

\item{MCS_runs}{The number of random samples drawn
(Monte-Carlo-Simulation runs)}

\item{rel_deviation}{Either one numeric value between 0 and 1 defining the
standard deviation relative to the measured value for all tracers or a vector
of the same length as 'tracer_names' defining a deviation for each tracer}
}
\value{
A data frame, columns corresponding to the tracers and rows correspoding to
the runs.
}
\description{
Random concentrations in sample based on measured data and assumed deviation
}
