% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_MCS_input}
\alias{normalize_MCS_input}
\title{Transforms tracer concentrations into normalized values between 0 and 1}
\usage{
normalize_MCS_input(MCS_input)
}
\arguments{
\item{MCS_input}{A list of sample Data frames created by
\code{\link[=get_all_sample_concentrations]{get_all_sample_concentrations()}}}
}
\value{
Same structured list as the input with normalized values instead of absolute
concentrations
}
\description{
Transforms tracer concentrations into normalized values between 0 and 1
}
\details{
Normalization per tracer by equation: \cr
(c_i - min(c)) / (max(c) - min(c)) \cr
This results in values between 0 (equals the overall minimal value of the
tracer within all samples) and 1 (equals the overall maximal value of the
tracer within all samples).
}
