% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_reshape.R
\name{reshape_table}
\alias{reshape_table}
\title{Reshape Data to Parameter-as-Columns-Structure}
\usage{
reshape_table(df_in, sampling_campaigns, bLOD = "na", aLOD = "na")
}
\arguments{
\item{df_in}{Data frame loaded with \code{\link[=load_lab_data]{load_lab_data()}}}

\item{sampling_campaigns}{One or more sampling Campaign IDs to be considered}

\item{bLOD}{Character string defining the handling of operator "<" (below
limit of detection). Either "half" for dividing the value by 2, "zero" or
"na" for setting the value to 0 or NA, respectively.}

\item{aLOD}{Character string defining the handling of operator ">" (above
limit of detection). Either "double" for multiplying the value by 2,
"use_limit" for using the limit value or "na" for setting the value to NA.}
}
\value{
A data frame of all data per sample in a row, arranged by the
starting time of the sample.
}
\description{
This function keeps the columns "sample_name", "sampling_campaign", "tBeg",
"tEnd" and adds one column per parameter. The operators are taken into
account.
}
